+++ sdk/face/ai-vision-face-rest/samples/v1-beta/javascript/livenessSessionWithVerify.js
@@ -2,7 +2,8 @@ const { randomUUID } = require("crypto");
 
 const { AzureKeyCredential } = require("@azure/core-auth");
 
-const createFaceClient = require("@azure-rest/ai-vision-face").default;
+const createFaceClient = require("@azure-rest/ai-vision-face").default,
+  { isUnexpected } = require("@azure-rest/ai-vision-face");
 
 /**
  * This sample demonstrates how to create a liveness detection session.
@@ -55,7 +56,7 @@ async function main() {
 
   // 2.Send a request to Face API to create a liveness detection session.
   const createLivenessSessionResponse = await client
-    .path("/detectLiveness/singleModal/sessions")
+    .path("/detectLivenessWithVerify/singleModal/sessions")
     .post({
       body: {
         livenessOperationMode: "Passive",
@@ -64,6 +65,9 @@ async function main() {
         authTokenTimeToLiveInSeconds: 60,
       },
     });
+  if (isUnexpected(createLivenessSessionResponse)) {
+    throw new Error(createLivenessSessionResponse.body.error.message);
+  }
   console.log("Create liveness session:");
   console.log(JSON.stringify(createLivenessSessionResponse.body, null, 2));
 
@@ -78,30 +82,42 @@ async function main() {
 
   // 8. After client devices perform the action, we can get the result from the following APIs.
   // Get session results.
-  const getLivenessSessionResponse = await client
-    .path("/detectLiveness/singleModal/sessions/{sessionId}", sessionId)
+  const getLivenessSessionResultResponse = await client
+    .path("/detectLivenessWithVerify/singleModal/sessions/{sessionId}", sessionId)
     .get();
+  if (isUnexpected(getLivenessSessionResultResponse)) {
+    throw new Error(getLivenessSessionResultResponse.body.error.message);
+  }
   console.log("Get liveness detection results:");
-  console.log(JSON.stringify(getLivenessSessionResponse.body, null, 2));
+  console.log(JSON.stringify(getLivenessSessionResultResponse.body, null, 2));
 
   // Get audit entries.
   const getAuditEntryResponse = await client
-    .path("/detectLiveness/singleModal/sessions/{sessionId}/audit", sessionId)
+    .path("/detectLivenessWithVerify/singleModal/sessions/{sessionId}/audit", sessionId)
     .get();
+  if (isUnexpected(getAuditEntryResponse)) {
+    throw new Error(getAuditEntryResponse.body.error.message);
+  }
   console.log("Get audit entries:");
   console.log(JSON.stringify(getAuditEntryResponse.body, null, 2));
 
   // We can also list all liveness sessions of this face account.
   const getLivenessSessionsResponse = await client
-    .path("/detectLiveness/singleModal/sessions")
+    .path("/detectLivenessWithVerify/singleModal/sessions")
     .get();
+  if (isUnexpected(getLivenessSessionsResponse)) {
+    throw new Error(getLivenessSessionsResponse.body.error.message);
+  }
   console.log("Get liveness sessions:");
   console.log(JSON.stringify(getLivenessSessionsResponse.body, null, 2));
 
   // Delete session.
   const deleteLivenessSessionResponse = await client
-    .path("/detectLiveness/singleModal/sessions/{sessionId}", sessionId)
+    .path("/detectLivenessWithVerify/singleModal/sessions/{sessionId}", sessionId)
     .delete();
+  if (isUnexpected(deleteLivenessSessionResponse)) {
+    throw new Error(deleteLivenessSessionResponse.body.error.message);
+  }
   console.log("Delete liveness session:");
   console.log(JSON.stringify(deleteLivenessSessionResponse.body, null, 2));
 }
